backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg[binary]==3.2.5
django-environ==0.11.2
dj-database-url==2.2.0
drf-spectacular==0.27.2